{"schema":"decodex/commit/1","summary":"Modularize live frame stream output","authority":"manual"}
